+++ src/main/java/wooteco/subway/controller/StationController.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -7,6 +7,10 @@ import org.springframework.web.bind.annotation.*;
 import java.net.URI;
 import java.util.List;
 import java.util.stream.Collectors;
+import wooteco.subway.dao.StationDao;
+import wooteco.subway.dto.StationRequest;
+import wooteco.subway.dto.StationResponse;
+import wooteco.subway.domain.station.Station;
 
 @RestController
 public class StationController {
