+++ src/main/java/wooteco/subway/domain/line/Line.java
@@ -1,0 +1,5 @@
+package wooteco.subway.domain.line;
+
+public class Line {
+
+}
