+++ src/main/java/wooteco/subway/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.dto;
 
 public class StationRequest {
     private String name;
